requirements.txt
Flask==2.3.3
Flask-Session==0.4.0
Werkzeug==2.3.7
standard-imghdr